Extensive changes to the parser and order entry features.
The parser was ported from a hand-coded LL(1) parser to the JParsec framework that allows for basic LL(k) parsing, as well as some extensions that allow for LL(infinity) parsing.  Both the lexer and the parser are constructed using the LL(k) parsing engine, to implement the non-trivial token parsing required in the Photon client.  The major change that the end user will see is that order commands in the "command entry area" must now be prefixed by "O " (for order).  For example "O B 100 IBM 123.4".

Additionally the StockOrderTicket has been reworked to implement a fairly generic form validation strategy.  Through the use of implementations of IControlValidator.java and IControlHighlighter.java, we have a way to ensure that user input is valid before passing it on to the internal controller.  For example an instance of the NumericTextValidator, and an instance of ParentColorHighlighter can be registered with a FormValidator to automatically validate numerical input, and provide user feedback in the case of an error.

git-svn-id: https://source.marketcetera.org/root/trunk@5739 501a4a96-914b-0410-8454-f5988a4ca596