Reverted the build.properties to create tarballs again
Modified the customTargets.xml to manually modify the linux tarballs to re-set the executable bit on the photon launcher (for some reason the bit is lost during packaging). this is done in the postBuild target.

Moved the code to manually modify contents of photon.ini for MacOS build to the postBuild target as well to streamline the process, avoiding doing it twice doing target assembly.

This should close the #91 bug


git-svn-id: https://source.marketcetera.org/root/trunk@6153 501a4a96-914b-0410-8454-f5988a4ca596