This is a fairly extensive refactoring of the UI code.  Specifically, we now make use of XSWT for building the UIs for order tickets.  We use JFace databinding wherever possible to bind the model to the user interface.
I think most of the important stuff now lives in the classes *OrderTicketModel *OrderTicketView and *OrderTicketController.  The class hierarchies are the same for model, view and controller.  For example the controller hierarchy has OrderTicketController at the root with two subclasses, StockOrderTicketController and OptionOrderTicketController.  If you can understand these 9 classes, I think that's most of it.

The controllers are mainly responsible for managing subscriptions to market data, and with interaction with other parts of the system.  For example the controllers are responsible for integrating with the Eclipse main preference store, such that the custom fields stored in there are available in the order ticket model and therefore view.  The model classes store data, and make heavy use of WritableValue and WritableList, which are provided by the databinding framework as observable data elements (that is they fire events when changed).

The data is bound to the UI using primarily calls to bindValue() .  These calls take 4 parameters: the target (UI) observable value, an event generating wrapper around the UI widget; the model observable (discussed above); a collection of "lambda functions" for the movement of data from the target to the model; and a similar collection of "lambda functions" for the model to target direction.  The lambda chain executes like this (in the model-to-target direction).

model --> after get IValidator --> IConverter --> after convert IValidator --> before set IValidator --> target

Any and all of the lambdas are optional.  The type of the output of one must match the input of the next.  There are two IValidators after the conversion, because the chain can be partially executed in some circumstances.

Take a look at the binding of the symbol text field to a custom observable value on the model.  It can be found in OrderTicketView#bindMessage() round about line 550 of that file.  This actually calls bindMessageValue() which is simply a helper function that stores the produced binding for later use.

git-svn-id: https://source.marketcetera.org/root/trunk@7132 501a4a96-914b-0410-8454-f5988a4ca596